chore(api): move organization-learner repo from shared to prescription context
