+++ api/src/prescription/learner-management/application/jobs/schedule-compute-organization-learners-certificability-job-controller.js
@@ -5,9 +5,9 @@ import { ComputeCertificabilityJob } from '../../../../prescription/learner-mana
 import { JobScheduleController } from '../../../../shared/application/jobs/job-schedule-controller.js';
 import { config } from '../../../../shared/config.js';
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
-import * as organizationLearnerRepository from '../../../../shared/infrastructure/repositories/organization-learner-repository.js';
 import { logger } from '../../../../shared/infrastructure/utils/logger.js';
 import { computeCertificabilityJobRepository } from '../../../learner-management/infrastructure/repositories/jobs/compute-certificability-job-repository.js';
+import * as organizationLearnerRepository from '../../../organization-learner/infrastructure/repositories/organization-learner-repository.js';
 import { ComputeOrganizationLearnerCertificabilityJobProvidedDateError } from '../../domain/errors.js';
 import { usecases } from '../../domain/usecases/index.js';
