+++ api/src/prescription/organization-place/domain/usecases/index.js
@@ -2,7 +2,7 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import * as organizationFeatureApi from '../../../../organizational-entities/application/api/organization-features-api.js';
-import * as organizationLearnerRepository from '../../../../shared/infrastructure/repositories/organization-learner-repository.js';
+import * as organizationLearnerRepository from '../../../../prescription/organization-learner/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationRepository from '../../../../shared/infrastructure/repositories/organization-repository.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
